requirements.txt
# Knowledge Graph
rdflib>=7.0.0
networkx>=3.0

# Embeddings and ML
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
Pillow>=9.0.0

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.28.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0